S07P31E102-94 #done #comment feat:jwt Notice CRUD (nested User Entity) done

See merge request s07-final/S07P31E102!18

+++ back/yourjs/src/main/java/com/project/yourjs/db/repository/NoticeRepository.java
@@ -4,6 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.project.yourjs.db.entity.Notice;
 
-public interface NoticeRepository extends JpaRepository<Notice, Long>{
-  
+public interface NoticeRepository extends JpaRepository<Notice, Integer>{
 }
